#  feat(postgres): use faster sql.DB instead of docker exec psql for snapshot/restore [rebased for main]