--- a/modules/postgres/postgres_test.go
+++ b/modules/postgres/postgres_test.go
@@ -26,38 +26,47 @@ const (
 	password = "password"
 )
 
+var (
+	// BasicWaitStrategies is a simple but reliable way to wait for postgres to start.
+	BasicWaitStrategies = testcontainers.WithWaitStrategy(
+		// First, we wait for the container to log readiness twice.
+		// This is because it will restart itself after the first startup.
+		wait.ForLog("database system is ready to accept connections").WithOccurrence(2),
+		// Then, we wait for docker to actually serve the port on localhost.
+		// For non-linux OSes like Mac and Windows, Docker or Rancher Desktop will have to start a separate proxy
+		// process that sometimes takes a little longer to initialize.
+		// Without this, the tests will be flaky on those OSes!
+		wait.ForListeningPort("5432/tcp"),
+	)
+)
+
 func TestPostgres(t *testing.T) {
 	ctx := context.Background()
 
 	tests := []struct {
 		name  string
 		image string
-		wait  wait.Strategy
 	}{
 		{
 			name:  "Postgres",
 			image: "docker.io/postgres:15.2-alpine",
-			wait:  wait.ForLog("database system is ready to accept connections").WithOccurrence(2).WithStartupTimeout(5 * time.Second),
 		},
 		{
 			name: "Timescale",
 			// timescale {
 			image: "docker.io/timescale/timescaledb:2.1.0-pg11",
-			wait:  wait.ForLog("database system is ready to accept connections").WithOccurrence(2).WithStartupTimeout(5 * time.Second),
 			// }
 		},
 		{
 			name: "Postgis",
 			// postgis {
 			image: "docker.io/postgis/postgis:12-3.0",
-			wait:  wait.ForLog("database system is ready to accept connections").WithOccurrence(2).WithStartupTimeout(30 * time.Second),
 			// }
 		},
 		{
 			name: "Pgvector",
 			// pgvector {
 			image: "docker.io/pgvector/pgvector:pg16",
-			wait:  wait.ForLog("database system is ready to accept connections").WithOccurrence(2).WithStartupTimeout(30 * time.Second),
 			// }
 		},
 	}
@@ -69,7 +78,7 @@ func TestPostgres(t *testing.T) {
 				postgres.WithDatabase(dbname),
 				postgres.WithUsername(user),
 				postgres.WithPassword(password),
-				testcontainers.WithWaitStrategy(tt.wait),
+				BasicWaitStrategies,
 			)
 			if err != nil {
 				t.Fatal(err)
@@ -166,7 +175,7 @@ func TestWithConfigFile(t *testing.T) {
 		postgres.WithDatabase(dbname),
 		postgres.WithUsername(user),
 		postgres.WithPassword(password),
-		testcontainers.WithWaitStrategy(wait.ForLog("database system is ready to accept connections").WithOccurrence(2).WithStartupTimeout(5*time.Second)),
+		BasicWaitStrategies,
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -197,7 +206,7 @@ func TestWithInitScript(t *testing.T) {
 		postgres.WithDatabase(dbname),
 		postgres.WithUsername(user),
 		postgres.WithPassword(password),
-		testcontainers.WithWaitStrategy(wait.ForLog("database system is ready to accept connections").WithOccurrence(2).WithStartupTimeout(5*time.Second)),
+		BasicWaitStrategies,
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -235,10 +244,7 @@ func TestSnapshot(t *testing.T) {
 		postgres.WithDatabase(dbname),
 		postgres.WithUsername(user),
 		postgres.WithPassword(password),
-		testcontainers.WithWaitStrategy(
-			wait.ForLog("database system is ready to accept connections").
-				WithOccurrence(2).
-				WithStartupTimeout(5*time.Second)),
+		BasicWaitStrategies,
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -341,10 +347,7 @@ func TestSnapshotWithOverrides(t *testing.T) {
 		postgres.WithDatabase(dbname),
 		postgres.WithUsername(user),
 		postgres.WithPassword(password),
-		testcontainers.WithWaitStrategy(
-			wait.ForLog("database system is ready to accept connections").
-				WithOccurrence(2).
-				WithStartupTimeout(5*time.Second)),
+		BasicWaitStrategies,
 	)
 	if err != nil {
 		t.Fatal(err)