See description of https://github.com/testcontainers/testcontainers-go/pull/2599

This is the same PR, but rebased for main.

The original description follows, just with latest benchmarks

--- 

<!-- Type of change
Please label this PR with one of the existing labels, depending on the scope of your change
-->

## What does this PR do?

The postgres module has functions for saving snapshots and restoring them later.
This feature is helpful for integration testing without having to manually clear the database.

This PR speeds up the snapshot/restore functionality by executing the necessary SQL commands via a regular `sql.DB` connection.

If the driver is not available in the test environment, or the commands fail for some other reason (SSL config, etc), it falls back to the old variant of calling `docker exec`.

## Why is it important?

`docker exec` can be unnecessarily slow, especially when using docker in a VM with Docker Desktop or colima.

On my laptop (MacBook Air M1), each call to `Container.Restore()` takes about 150~200ms.
In a small subset of 20 of our tests, where each test restores the database before starting, with creating the initial container, this brings the total time up to 12 seconds, which is quite a lot for a few small database tests.

After these changes, the same subset of tests now takes 6 seconds.

## Related issues

- Relates to https://github.com/testcontainers/testcontainers-go/pull/2199
- Relates to https://github.com/testcontainers/testcontainers-go/pull/2599

## How to test this PR

```go
cd modules/postgres
go test ./...
```

## Follow-ups

I've also considered caching the DB pool that is created to manage the databases.
After running some benchmark with this code (from branch X), I noticed it does not have a big impact on the whole test suite - my internal test suite does not change in execution time (since the time is dominated by re-connecting the actual test client, which will be terminated when the database gets deleted during the restore).
The module/postgres test suite even gets slower:

cached: https://github.com/testcontainers/testcontainers-go/compare/main...cfstras:postgres-snapshot-sql-cached-rebased-main
not cached: this PR



<details>
  <summary>testcontainers main, sql, not cached: <b>42.227 s ±  0.635</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     42.227 s ±  0.635 s    [User: 1.213 s, System: 1.034 s]
  Range (min … max):   41.832 s … 44.006 s    10 runs
</pre>
</details>

<details>
  <summary>testcontainers main, sql, cached: <b>43.207 s ±  1.549</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     43.207 s ±  1.549 s    [User: 1.247 s, System: 1.057 s]
  Range (min … max):   41.745 s … 45.664 s    10 runs
</pre>
</details>

<details>
  <summary>testcontainers main, without PR (but amended tests): <b>45.298 s ±  5.610 s</b></summary>
Branch: <a href="https://github.com/testcontainers/testcontainers-go/compare/main...cfstras:testcontainers-go:main-with-postgres-sql-test-fixes">cfstras:testcontainers-go:main-with-postgres-sql-test-fixes</a>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     45.298 s ±  5.610 s    [User: 1.218 s, System: 1.005 s]
  Range (min … max):   41.243 s … 57.096 s    10 runs
</pre>
</details>

<details>
  <summary>internal testsuite, with testcontainers main, sql, cached, with ryuk: <b>14.123 s ±  1.795 s</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     14.123 s ±  1.795 s    [User: 9.589 s, System: 2.749 s]
  Range (min … max):   11.407 s … 16.748 s    10 runs
</pre>
</details>

<details>
  <summary>internal testsuite, with testcontainers main, sql, cached, without ryuk: <b>11.299 s ±  1.503 s</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     11.299 s ±  1.503 s    [User: 9.818 s, System: 2.728 s]
  Range (min … max):    9.928 s … 13.781 s    10 runs
</pre>
</details>

<details>
  <summary>internal testsuite, with testcontainers main, sql, not cached, without ryuk: <b>10.597 s ±  0.641 s</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     10.597 s ±  0.641 s    [User: 9.812 s, System: 2.820 s]
  Range (min … max):    9.894 s … 12.029 s    10 runs
</pre>
</details>

<details>
  <summary>internal testsuite, with testcontainers main, without my PR, not cached, without ryuk: <b>14.245 s ±  0.639 s</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     14.245 s ±  0.639 s    [User: 9.671 s, System: 2.760 s]
  Range (min … max):   13.742 s … 15.980 s    10 runs
</pre>
</details>

<details>
  <summary>internal testsuite, with testcontainers main, docker-exec fallback, not cached, without ryuk: <b>14.216 s ±  0.450 s</b></summary>
<pre>
➜ hyperfine "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     14.216 s ±  0.450 s    [User: 9.566 s, System: 2.654 s]
  Range (min … max):   13.735 s … 15.278 s    10 runs
</pre>
</details>