@mdelapenya 2024-06-28
@cfstras @ash2k I went ahead and added https://github.com/testcontainers/testcontainers-go/pull/2600/commits/d8468eca864b3662daef6cd5e200916f18b40b7b converting the global variable into a container option. PLMK what you think.

Cheers!